Upgrades jsdom to latest and a bunch of other dep upgrades
This upgrades to jsdom@8.5.0 and corrects some descrepencies in
the tests that the upgrade introduced. It also upgrades other
dependencies to their latest.

Except for grunt... for now I am leaving that one alone.

Also upgrades the eslint file to the latest version and created a
.eslintignore file to remove the /tmp and /dist directories from
the eslint checks. eslint is much happier now.

Lastly swapped a switch statement for a simple ternary in t_call
because it reads a little easier.