generate index for use in Canvas
== scope-preservation

the index file contains a flat list of all the phrases and the scopes
they are defined in, and the scopes they are used in. This is used to
preserve the keys by avoiding the ambiguity introduced when we produce
them as a tree:

    {
      "a": {
        "b": {
          "one": "One banana",
          "other": "Many bananas"
        }
      }
    }

it's hard to reliably tell that "one.two" is a pluralized phrase and
that "one" is the scope for "one.two" and that "one.two" is NOT a
scope for "one.two.one". With the index, the scope is preserved, so we
don't even to differentiate and this is no longer a problem:

    [
      { "key": "a.b.one", "scope": "a.b" },
      { "key": "a.b.other", "scope": "a.b" },
    ]

== monkey patching reduction

i18nliner-canvas no longer monkey-patches the upstream Commands or
Processor objects, instead it provides its own definitions that
inherit from the core ones and the core code now supports using those
custom definitions

== behavior changes

absolute keys are no longer allowed when they are unscoped... this
seems to have been an oversight of the initial implementation as there
was one occurrence of it in Canvas and the rest of the code does not
allow for unscoped keys